DOCS-1898 - Refactor code
- Unify the structure of responses
- Split functions into smaller units
- Handle exceptions

+++ server/controllers/buildsController.js
@@ -2,36 +2,63 @@ const { Client } = require('@elastic/elasticsearch');
 const elasticClient = new Client({ node: process.env.CONFIG_DB_URL });
 const buildsIndexName = 'builds';
 
-async function getBuilds(buildId) {
+async function getAllBuilds() {
   try {
-    let queryBody;
-    if (buildId) {
-      queryBody = {
+    const response = await elasticClient.search({
+      index: buildsIndexName,
+      size: 10000,
+      body: {
+        query: {
+          match_all: {},
+        },
+      },
+    });
+    const hits = response.body.hits.hits.map(h => ({
+      _id: h._id,
+      build_id: h._source.build_id,
+      resources: h._source.resources,
+    }));
+    return {
+      body: hits,
+      status: response.statusCode,
+    };
+  } catch (err) {
+    console.error(err);
+    return err.message;
+  }
+}
+
+async function getBuildById(buildId) {
+  try {
+    const response = await elasticClient.search({
+      index: buildsIndexName,
+      size: 10000,
+      body: {
         query: {
           match: {
             build_id: buildId,
           },
         },
-      };
-    } else {
-      queryBody = {
-        query: {
-          match_all: {},
-        },
-      };
-    }
-
-    const response = await elasticClient.search({
-      index: buildsIndexName,
-      size: 10000,
-      body: queryBody,
+      },
     });
-    const hits = response.body.hits.hits;
-    return hits.map(h => ({
+    const hit = response.body.hits.hits.map(h => ({
       _id: h._id,
       build_id: h._source.build_id,
       resources: h._source.resources,
-    }));
+    }))[0];
+    if (hit) {
+      return {
+        body: hit,
+        status: response.statusCode,
+      };
+    } else {
+      return {
+        body: {
+          message: `Build with ID "${buildId}" not found`,
+        },
+        status: 404,
+      };
+    }
   } catch (err) {
     console.error(err);
     return err.message;
@@ -44,14 +71,17 @@ async function addOrUpdateBuild(reqBody) {
       index: buildsIndexName,
       body: reqBody,
     };
-    const buildsInfo = await getBuilds(reqBody.build_id);
-    if (buildsInfo.length > 0) {
-      requestParams.id = buildsInfo[0]._id;
+    const buildInfo = await getBuildById(reqBody.build_id);
+    const buildUniqueId = buildInfo.body._id;
+    if (buildUniqueId) {
+      requestParams.id = buildUniqueId;
     }
     const response = await elasticClient.index(requestParams);
     return {
-      _id: response.body._id,
-      result: response.body.result,
+      body: {
+        _id: response.body._id,
+        result: response.body.result,
+      },
       status: response.statusCode,
     };
   } catch (err) {
@@ -62,16 +92,23 @@ async function addOrUpdateBuild(reqBody) {
 
 async function deleteBuild(buildId) {
   try {
-    const buildsInfo = await getBuilds(buildId);
-    const response = await elasticClient.delete({
-      index: buildsIndexName,
-      id: buildsInfo[0]._id,
-    });
-    return {
-      _id: response.body._id,
-      result: response.body.result,
-      status: response.statusCode,
-    };
+    const buildInfo = await getBuildById(buildId);
+    const buildUniqueId = buildInfo.body._id;
+    if (buildUniqueId) {
+      const response = await elasticClient.delete({
+        index: buildsIndexName,
+        id: buildUniqueId,
+      });
+      return {
+        body: {
+          _id: response.body._id,
+          result: response.body.result,
+        },
+        status: response.statusCode,
+      };
+    } else {
+      return buildInfo;
+    }
   } catch (err) {
     console.error(err);
     return err.message;
@@ -79,7 +116,8 @@ async function deleteBuild(buildId) {
 }
 
 module.exports = {
-  getBuilds,
+  getAllBuilds,
+  getBuildById,
   addOrUpdateBuild,
   deleteBuild,
 };
